backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
pyjwt==2.8.0
psycopg2-binary==2.9.9
requests==2.31.0
